Move external urls, add missing DocumentTitles
