+++ src/client/config/app.ts
@@ -10,6 +10,7 @@ export const PLACEHOLDER_IMAGE_URL =
   'data:image/gif;base64,R0lGODlhAQABAIAAAMLCwgAAACH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==';
 
 // Legacy export, these used to live in this file
-export { ExternalUrls, Colors } from '../../universal/config/feature-toggles';
+export { Colors } from '../../universal/config/feature-toggles';
+export { ExternalUrls } from './external-urls';
 
 export const MAX_TABLE_ROWS_ON_THEMA_PAGINA = 3;
